third_party/WebKit/LayoutTests/platform/mac/external/wpt/IndexedDB/idbobjectstore-rename-store-expected.txt

This is a testharness.js-based test.
FAIL IndexedDB object store rename in new transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename in the transaction where it is created assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename covers index assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename covers key generator assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename to the same name succeeds assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename to the name of a deleted store succeeds assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store swapping via renames succeeds assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store rename stringifies non-string names assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store can be renamed to "" assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store can be renamed to "\u0000" assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB object store can be renamed to "\uDC00\uD800" assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
